make miri schemas work with asdf and datamdoels 0.13.4
