+++ datamodels/schemas/miri_bad_pixel_mask.schema.yaml
@@ -2,5 +2,5 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/mask.schema
