+++ datamodels/schemas/miri_colourcorrection_imaging.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI imaging black body colour correction model
   properties:
