+++ datamodels/schemas/miri_colourcorrection_powerlaw.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI imaging power law colour correction model
   properties:
