+++ datamodels/schemas/miri_core.schema.yaml
@@ -3,5 +3,7 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/referencefile.schema
+
+$schema: http://stsci.edu/schemas/fits-schema/fits-schema
