+++ datamodels/schemas/miri_dark_reference.schema.yaml
@@ -2,11 +2,11 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - type: object
   properties:
     meta:
-      $ref: miri_metadata_dataerr.schema
+      $ref: miri_metadata_dataerr.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/darkMIRI.schema
 - type: object
   title: MIRI DARK reference model
