+++ datamodels/schemas/miri_distortion_mrs34_CDP6.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI MRS Distortion Maps
   properties:
