+++ datamodels/schemas/miri_fluxconversion_imaging.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI imaging flux conversion model
   properties:
