+++ datamodels/schemas/miri_fluxconversion_mrs.schema.yaml
@@ -1,6 +1,6 @@
 allOf:
-- $ref: miri_measured.schema
-- $ref: keyword_photuja2.schema
+- $ref: miri_measured.schema.yaml
+- $ref: keyword_photuja2.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/keyword_photmjsr.schema
 - type: object
   title: MIRI MRS flux conversion model
