+++ datamodels/schemas/miri_fringe_frequencies_mrs.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI MRS fringe frequencies model
   properties:
