+++ datamodels/schemas/miri_jump.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI ramp jump correction model
   properties:
