+++ datamodels/schemas/miri_photom.schema.yaml
@@ -2,8 +2,8 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
-- $ref: mirimg_photom.schema
+      $ref: miri_metadata.schema.yaml
+- $ref: mirimg_photom.schema.yaml
 - type: object
   title: MIRI photometric flux conversion model
   properties:
