+++ datamodels/schemas/miri_photom_CDP5.schema.yaml
@@ -2,7 +2,7 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/mirimg_photom.schema
 - type: object
   title: MIRI photometric flux conversion model
