+++ datamodels/schemas/miri_pixel_saturation.schema.yaml
@@ -2,11 +2,11 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - type: object
   properties:
     meta:
-      $ref: miri_metadata_dataerr.schema
+      $ref: miri_metadata_dataerr.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/saturation.schema
 - type: object
   title: MIRI pixel saturation calibration data
