+++ datamodels/schemas/miri_psf.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_measured.schema
+- $ref: miri_measured.schema.yaml
 - type: object
   title: JWST/MIRI point spread function
   properties:
