+++ datamodels/schemas/miri_psf_lrs.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_psf.schema
+- $ref: miri_psf.schema.yaml
 - type: object
   title: JWST/MIRI LRS point spread function
   properties:
