+++ datamodels/schemas/miri_reset.schema.yaml
@@ -2,9 +2,9 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - type: object
   properties:
     meta:
-      $ref: miri_metadata_dataerr.schema
+      $ref: miri_metadata_dataerr.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/reset.schema
