+++ datamodels/schemas/miri_reset_switch_charge_decay_CDP6.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI detector reset switch charge decay model
   properties:
