+++ datamodels/schemas/miri_simple.schema.yaml
@@ -1,23 +1,6 @@
 allOf:
-- $ref: miri_core.schema
-- $ref: http://stsci.edu/schemas/jwst_datamodel/wcsinfo.schema
 - type: object
-  title: Basic MIRI data
   properties:
     meta:
-      data:
-        type: object
-        title: Information about the data array
-        properties:
-          units:
-            type: string
-            title: Units of the data array
-            fits_hdu: SCI
-            fits_keyword: BUNIT
-    data:
-      title: The science data
-      fits_hdu: SCI
-      default: 0
-      units: ''
-      axes: [hypercubes, cubes, planes, rows, columns]
-      datatype: float32
+      $ref: miri_metadata.schema.yaml
+- $ref: http://stsci.edu/schemas/jwst_datamodel/image.schema
