+++ datamodels/schemas/miri_spectral_spatial_resolution_mrs.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI MRS spectral and spatial resolution model
   properties:
