+++ datamodels/schemas/miri_straylight_mrs.schema.yaml
@@ -2,6 +2,6 @@ allOf:
 - type: object
   properties:
     meta:
-      $ref: miri_metadata.schema
+      $ref: miri_metadata.schema.yaml
 - $ref: http://stsci.edu/schemas/jwst_datamodel/straylight.schema
