+++ datamodels/schemas/miri_straylight_mrs_CDP3.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_bad_pixel_mask.schema
+- $ref: miri_bad_pixel_mask.schema.yaml
 - type: object
   title: MIRI Old Format MRS Straylight pixel mask
   properties:
