+++ datamodels/schemas/miri_telescope_emission.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_measured.schema
+- $ref: miri_measured.schema.yaml
 - type: object
   title: JWST/MIRI telescope emission model
   properties:
