+++ datamodels/schemas/miri_transmission_correction_mrs.schema.yaml
@@ -1,5 +1,5 @@
 allOf:
-- $ref: miri_core.schema
+- $ref: miri_core.schema.yaml
 - type: object
   title: MIRI MRS transmission correction model
   properties:
